refactor: merged Footer and Header component to +layout.svelte
