--- sites/svelte.water-gis.com/src/lib/components/Footer.svelte
@@ -1,18 +1,0 @@
-<footer class="footer">
-	<div class="content has-text-centered">
-		<p>
-			<strong><a href="https://github.com/watergis" target="_blank">GIS for Water</a></strong> by
-			<a href="https://github.com/JinIgarashi" target="_blank">Jin Igarashi</a>. The source code is
-			licensed
-			<a href="https://opensource.org/licenses/mit-license.php" target="_blank">MIT</a>. The website
-			content is licensed
-			<a href="https://creativecommons.org/licenses/by-nc-sa/4.0/" target="_blank"
-				>CC BY NC SA 4.0</a
-			>.
-		</p>
-	</div>
-</footer>
-
-<style lang="scss">
-	@import 'bulma/bulma.sass';
-</style>
